refactor: streamline package layout
