+++ src/client/masonry/resolveMasonryProps.ts
@@ -1,4 +1,4 @@
-import { resolveCommonProps, resolveResponsiveParameter } from "../../core/utils";
+import { resolveCommonProps, resolveResponsiveParameter } from "../../utils";
 import { MasonryPhotoAlbumProps, Photo } from "../../types";
 
 export default function resolveMasonryProps<TPhoto extends Photo>(
