+++ src/scroll/InfiniteScroll.tsx
@@ -1,10 +1,10 @@
 import type React from "react";
 import { Children, cloneElement, isValidElement, useCallback, useRef, useState } from "react";
 
-import useEventCallback from "./useEventCallback";
-import { CommonPhotoAlbumProps, Photo, RenderTrackProps } from "../types";
 import Offscreen from "./Offscreen";
+import useEventCallback from "./useEventCallback";
 import useIntersectionObserver from "./useIntersectionObserver";
+import { CommonPhotoAlbumProps, Photo, RenderTrackProps } from "../types";
 
 enum Status {
   IDLE,
