+++ src/utils/ratio.ts
@@ -1,4 +1,4 @@
-import { Photo } from "../../types";
+import { Photo } from "../types";
 
 export default function ratio({ width, height }: Pick<Photo, "width" | "height">) {
   return width / height;
