+++ vite.config.ts
@@ -20,8 +20,8 @@ export default defineConfig({
     lib: {
       entry: {
         index: "src/index.ts",
-        "core/utils": "src/core/utils/index.ts",
-        "core/static": "src/core/static/index.ts",
+        "utils/index": "src/utils/index.ts",
+        "static/index": "src/static/index.ts",
         "layouts/rows": "src/layouts/rows/index.ts",
         "layouts/columns": "src/layouts/columns/index.ts",
         "layouts/masonry": "src/layouts/masonry/index.ts",
